[C] Consolidate slate utils
