+++ client/src/global/components/form/ContentEditor/components/controls/ToolbarRTE.js
@@ -15,7 +15,7 @@ import {
 } from "reakit/Menu";
 import { Node } from "slate";
 import { useSlateStatic, useSlateSelection } from "slate-react";
-import { getAncestors } from "../../transforms/utils";
+import { getAncestors } from "../../utils/slate";
 import { HtmlBreadcrumbsContext } from "../../index";
 import { rteElements, inlineNodes } from "../../utils/elements";
 import IconComposer from "global/components/utility/IconComposer";
