+++ client/src/global/components/form/ContentEditor/components/controls/buttons/LinkButton.js
@@ -1,6 +1,6 @@
 import React, { forwardRef, useRef } from "react";
 import { useTranslation } from "react-i18next";
-import { Editor, Element as SlateElement, Transforms, Range } from "slate";
+import { Editor, Range } from "slate";
 import { useSlate, ReactEditor } from "slate-react";
 import Utility from "global/components/utility";
 import InsertLinkForm from "./insert/LinkForm";
@@ -10,43 +10,8 @@ import Tooltip from "global/components/atomic/Tooltip";
 import TooltipContent from "./TooltipContent";
 import { hotkeys, labels } from "./TooltipContent/hotkeys";
 import { onModalClose } from "./utils";
+import { isLinkActive, wrapLink, unwrapLink } from "../../../utils/slate";
 import * as Styled from "./styles";
-
-export const isLinkActive = editor => {
-  const [link] = Editor.nodes(editor, {
-    match: n =>
-      !Editor.isEditor(n) && SlateElement.isElement(n) && n.type === "a"
-  });
-  return !!link;
-};
-
-export const unwrapLink = editor => {
-  Transforms.unwrapNodes(editor, {
-    match: n =>
-      !Editor.isEditor(n) && SlateElement.isElement(n) && n.type === "a"
-  });
-  ReactEditor.focus(editor);
-};
-
-export const wrapLink = (editor, url, text) => {
-  const { selection } = editor ?? {};
-  if (!selection) return;
-  const isCollapsed = selection && Range.isCollapsed(selection);
-  const link = {
-    type: "a",
-    htmlAttrs: { href: url },
-    children: isCollapsed ? [{ text: text ?? url }] : []
-  };
-
-  if (isCollapsed) {
-    Transforms.insertNodes(editor, link);
-  } else {
-    Transforms.wrapNodes(editor, link, { split: true });
-    Transforms.collapse(editor, { edge: "end" });
-  }
-
-  ReactEditor.focus(editor);
-};
 
 const LinkButton = ({ icon, size, ...rest }, ref) => {
   const editor = useSlate();
