+++ client/src/global/components/form/ContentEditor/components/controls/buttons/utils.js
@@ -1,7 +1,5 @@
-import { Transforms, Editor } from "slate";
+import { Transforms } from "slate";
 import { ReactEditor } from "slate-react";
-import { isBlockActive } from "./BlockButton";
-import { containers } from "../../../utils/elements";
 
 export const onModalClose = (editor, selection) => close => {
   close();
@@ -12,27 +10,4 @@ export const onModalClose = (editor, selection) => close => {
     Transforms.select(editor, selectionToUse);
     ReactEditor.focus(editor);
   }, 200);
-};
-
-export const getNearestContainer = (editor, selection) => {
-  const [container, containerPath] = Editor.above(editor, {
-    at: Editor.unhangRange(editor, selection),
-    match: n =>
-      !Editor.isEditor(n) &&
-      containers.includes(n.type) &&
-      !n.nodeName &&
-      !n.slateOnly
-  });
-  return { container, containerPath };
-};
-
-export const getActiveBlock = (editor, opts) => {
-  const active = opts
-    .map(o => [o, isBlockActive(editor, o)])
-    .reduce((obj, o) => ({ ...obj, [o[0]]: o[1] }), {});
-  const activeCount = opts.map(o => isBlockActive(editor, o)).filter(Boolean)
-    .length;
-
-  if (activeCount !== 1) return "";
-  return Object.keys(active).find(o => active[o]);
 };
