--- client/src/global/components/form/ContentEditor/transforms/utils.js
@@ -1,124 +1,0 @@
-import { Transforms, Editor, Node } from "slate";
-import { ReactEditor } from "slate-react";
-import has from "lodash/has";
-
-export const setSelectionAtPoint = (editor, path, offset = 0) => {
-  Transforms.select(editor, {
-    anchor: { path, offset },
-    focus: { path, offset }
-  });
-};
-
-export const getListItemNode = (editor, path) =>
-  Editor.above(editor, {
-    at: path,
-    match: n => n.type === "li"
-  }) ?? [];
-
-export const getListNode = (editor, path) =>
-  Editor.above(editor, {
-    at: path,
-    match: n => n.type === "ul" || n.type === "ol"
-  }) ?? [];
-
-export const findListChild = iterator => {
-  const { value } = iterator.next();
-  if (!value) return [];
-  const [node, path] = value;
-  if (node.type === "ul" || node.type === "ol") return [node, path];
-  return findListChild(iterator);
-};
-
-export const getTextContent = (editor, iterator, str = "") => {
-  const { value } = iterator.next();
-  if (!value) return str;
-  const [node] = value;
-  const nextStr = node.text ? str + node.text : str;
-  return getTextContent(editor, iterator, nextStr);
-};
-
-export const getTextLocation = (editor, iterator, path, str = "") => {
-  const { value } = iterator.next();
-  if (!value) return { textContent: str, lastPath: path };
-  const [node, nodePath] = value;
-  if (node.type === "list-sibling")
-    return getTextLocation(editor, Node.children(editor, nodePath), nodePath);
-  const isTextNode = has(node, "text");
-  const nextStr = isTextNode ? node.text : str;
-  const nextPath = isTextNode ? nodePath : path;
-  return getTextLocation(editor, iterator, nextPath, nextStr);
-};
-
-export const getIndentSelectionLocation = (editor, path) => {
-  const textIterator = Node.children(editor, path);
-  const { textContent, lastPath } = getTextLocation(editor, textIterator, path);
-  return { path: lastPath, offset: textContent.length };
-};
-
-export const nestListItem = ({ editor, srcPath, destPath, node, type }) => {
-  const children = Array.isArray(node) ? node : [node];
-  Transforms.removeNodes(editor, { at: srcPath });
-  Transforms.insertNodes(editor, [{ type, children }], {
-    at: destPath
-  });
-};
-
-export const liftListItem = ({ editor, srcPath, destPath, unwrap = true }) => {
-  if (destPath) Transforms.moveNodes(editor, { at: srcPath, to: destPath });
-  if (unwrap) {
-    const targetPath = destPath ?? srcPath;
-    Transforms.unwrapNodes(editor, { at: targetPath });
-  }
-};
-
-export const handleInlineItemChildren = ({ editor, at, nodes }) => {
-  if (nodes)
-    return Transforms.insertNodes(
-      editor,
-      [{ type: "list-sibling", slateOnly: true, children: nodes }],
-      { at }
-    );
-  return Transforms.unwrapNodes(editor, {
-    at,
-    match: n => n.type === "list-sibling"
-  });
-};
-
-export const getCommonBlock = editor => {
-  const range = Editor.unhangRange(editor, editor.selection, { voids: true });
-
-  const [common, path] = Node.common(
-    editor,
-    range.anchor.path,
-    range.focus.path
-  );
-
-  if (Editor.isBlock(editor, common) || Editor.isEditor(common)) {
-    return [common, path];
-  } else {
-    return Editor.above(editor, {
-      at: path,
-      match: n => Editor.isBlock(editor, n) || Editor.isEditor(n)
-    });
-  }
-};
-
-export const formatNodeLabel = node => {
-  let label = node.type === "void" ? node.nodeName : node.type;
-  const { htmlAttrs } = node ?? {};
-  if (htmlAttrs?.id) label = `${label}#${htmlAttrs.id}`;
-  if (htmlAttrs?.class) label = `${label}.${htmlAttrs.class.replace(" ", ".")}`;
-  return label;
-};
-
-export const getAncestors = (editor, iterator, list = {}) => {
-  const { value } = iterator.next();
-  if (!value) return list;
-  const [node] = value;
-  const key = ReactEditor.findKey(editor, node);
-  const next =
-    node.type && !node.slateOnly
-      ? { ...list, [key.id]: formatNodeLabel(node) }
-      : list;
-  return getAncestors(editor, iterator, next);
-};
